% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon_bar.R
\name{moon_bar}
\alias{moon_bar}
\title{Pre-designed Bar Charts}
\usage{
moon_bar(
  data,
  groups,
  fill_by = NULL,
  wrap_by = NULL,
  palette = NULL,
  text_mode = NULL
)
}
\arguments{
\item{data}{dataframe or tibble. Dataset to be used for the plot.}

\item{groups}{character vector. Defines the variable(s) that should be used for frequency calculations.
This function does not accept vectors with more than 2 variables, as bar charts would become unreadable
with more than 3 attributes (2 groups + frequency).}

\item{fill_by}{character. When using 2 groups, a method should always be chosen to insert the second group into the
plot. This parameter defines the which variable should be used as the fill color for the chart.}

\item{wrap_by}{character. When using 2 groups, a method should always be chosen to insert the second group into the
plot. This parameter defines the which variable should be used to separate (wrap) the charts.}

\item{palette}{character/vector. Pick a color to be used on the chart. Keep in mind that the number of colors
specified should be equal to the number of categories in the grouping variable.}

\item{text_mode}{character. Pick a method to display the frequency labels. Accepts the values: \code{"n"}, for
frequency, \code{"p"}, for the frequency in percentage and \code{"np"} or \code{"pn"} for frequency in absolute value and relative percentage.}
}
\value{
the requested plot.
}
\description{
Creates bar charts with the appropriate groups, frequencies, methods, themes and other features
according to the Moon theme definitions.
}
